+++ NEWS.md
@@ -1,5 +1,6 @@
 # git head
 
+* Add a workaround for `EnsDb`/`BSGenome` mismatch on `UCSC` names (#3)
 * Fix a bug where multi-exon dropping could subset incorrectly (#5)
 * Re-introduce filter that CDS width must be equal variant width (#5)